Feature - Select player to add cards on cheat mode.
I was testing (I mean on "real game", not unit testing). And I found i was very difficult to test vs Bot. So, I add a kind of selector for choosing the player who will get a card (It's only working for adding card method).

With this cheat I can test it very fast.